refactor(reducers): move `theme`
