+++ app/renderer/src/reducers/theme/index.js
@@ -1,0 +1,1 @@
+export { default } from './theme';
